probe_all now probes Neurons and LearningRules
In order to wrap my head around how to do this, I also simplified
the probe_all function in general. One major difference between
the old and new implementations is that in the new implementation,
all probes are added to the top-level network rather than the network
in which the probed object resides. I've found that this largely
mirrors how I've done probing in my models, but this can be reverted.