Fixed spasafe_names function to ignore additional non-SPA characters
Implement a more general spasafe_name function that removes all invalid
symbols with a regex instead of doing a replace for every single
symbol. Spaces and hyphens will be converted to underscores (except for
spaces at the beginning or end of the name). Also added tests for these
things including a simple test case for the spasafe_names function.

This changes might not be strictly required for the use case here, but
I have come across other uses cases where I needed such a function, so
it might be good to be more general. It is likely that I will use this
implementation to provide such a function in the new nengo_spa
implementation.

Also added documentation.