Remove the SocketCloseThread.
This simplifies the code a lot. It also means that the sockets are not
closed until all references to them are gone. To avoid problems with
the OS complaining about addresses still being in use, we also allow
the reuse of addresses. This means that the user might have to ensure
that no other program is using the socket.